+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJPZM9R8ZQW1FPK87P1XBYNE"
+        "01KJQ4GJ40DCH2VJ3TPPCXWX2K"
     )
     r"""The unique identifier of the data source"""