+++ packages/orq-rc/src/orq_ai_sdk/models/createidentityop.py
@@ -137,7 +137,7 @@ class CreateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-02T09:56:02.262Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-02T11:21:22.694Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")