+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJPZM9R7APPMNYBPNKQYRC7X"
+        "01KJQ4GJ3ZJRRDKEYP16NXRZXK"
     )
     r"""The unique identifier of the data source"""