+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJPZM9RBGRGBXSMM5P36DTD6"
+        "01KJQ4GJ4367K4GNVDBM2EHZJ6"
     )
     r"""The unique identifier of the data source"""